+++ xrlint/_linter/rulectx.py
@@ -17,6 +17,7 @@ class RuleContextImpl(RuleContext):
         dataset: xr.Dataset,
         file_path: str,
         file_index: int | None,
+        opening_time: float | None,
     ):
         assert config is not None
         assert dataset is not None
@@ -26,6 +27,7 @@ class RuleContextImpl(RuleContext):
         self._dataset = dataset
         self._file_path = file_path
         self._file_index = file_index
+        self._opening_time = opening_time
         self.messages: list[Message] = []
         self.rule_id: str | None = None
         self.severity: Literal[1, 2] = SEVERITY_ERROR
@@ -50,6 +52,10 @@ class RuleContextImpl(RuleContext):
     @property
     def file_index(self) -> int | None:
         return self._file_index
+
+    @property
+    def opening_time(self) -> float | None:
+        return self._opening_time
 
     def report(
         self,