+++ xrlint/_linter/validate.py
@@ -1,3 +1,4 @@
+import time
 from typing import Any
 
 import xarray as xr
@@ -15,7 +16,7 @@ def validate_dataset(config_obj: ConfigObject, dataset: Any, file_path: str):
     assert dataset is not None
     assert isinstance(file_path, str)
     if isinstance(dataset, xr.Dataset):
-        messages = _validate_dataset(config_obj, dataset, file_path, None)
+        messages = _validate_dataset(config_obj, dataset, file_path, None, None)
     else:
         messages = _open_and_validate_dataset(config_obj, dataset, file_path)
     return Result.new(config_object=config_obj, messages=messages, file_path=file_path)
@@ -26,12 +27,13 @@ def _validate_dataset(
     dataset: xr.Dataset,
     file_path: str,
     file_index: int | None,
+    opening_time: float | None,
 ) -> list[Message]:
     assert isinstance(config_obj, ConfigObject)
     assert isinstance(dataset, xr.Dataset)
     assert isinstance(file_path, str)
 
-    context = RuleContextImpl(config_obj, dataset, file_path, file_index)
+    context = RuleContextImpl(config_obj, dataset, file_path, file_index, opening_time)
     for rule_id, rule_config in config_obj.rules.items():
         with context.use_state(rule_id=rule_id):
             apply_rule(context, rule_id, rule_config)
@@ -48,24 +50,28 @@ def _open_and_validate_dataset(
     opener_options = config_obj.opener_options or {}
     if config_obj.processor is not None:
         processor_op = config_obj.get_processor_op(config_obj.processor)
+        t0 = time.time()
         try:
             ds_path_list = processor_op.preprocess(file_path, opener_options)
         except (OSError, ValueError, TypeError) as e:
             return [new_fatal_message(str(e))]
+        opening_time = time.time() - t0
         return processor_op.postprocess(
             [
-                _validate_dataset(config_obj, ds, path, i)
+                _validate_dataset(config_obj, ds, path, i, opening_time)
                 for i, (ds, path) in enumerate(ds_path_list)
             ],
             file_path,
         )
     else:
+        t0 = time.time()
         try:
             dataset = _open_dataset(ds_source, opener_options, file_path)
         except (OSError, ValueError, TypeError) as e:
             return [new_fatal_message(str(e))]
+        opening_time = time.time() - t0
         with dataset:
-            return _validate_dataset(config_obj, dataset, file_path, None)
+            return _validate_dataset(config_obj, dataset, file_path, None, opening_time)
 
 
 def _open_dataset(