+++ CHANGES.md
@@ -4,6 +4,9 @@
 
 ### Adjustments and Enhancements
 
+- Added a new core rule `opening-time` that can be used to check the
+  time it takes to open datasets.
+
 - Added HTML styling for both CLI output (`--format html`) and rendering
   of `Result` objects in Jupyter notebooks.